Move all .js files to .ts/.tsx.
